HW1/pyproject.toml
[tool.poetry]
name = "mlops-wine"
version = "0.1.0"
description = ""
authors = ["Alexander Skovorodko <alexanderskovorodko@gmail.com>"]
readme = "README.md"
packages = [{include = "mlops_wine"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.2"
pandas = "^1.5.3"
flask-restx = "^1.0.6"
python-dotenv = "^0.21.1"
scikit-learn = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
